requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# python-steganography/compare.py: 1
# python-steganography/stegano.py: 1
Pillow == 7.0.0

# python-steganography/compare.py: 2
# python-steganography/stegano.py: 2
numpy == 1.17.4
